optimization/requirements.txt
transformers
optimum[onnxruntime]
onnxruntime
torch
Pillow
numpy
requests
